+++ application/controls/finder-box-control/finder-box-control.js
@@ -1,0 +1,9 @@
+/**
+ * Copyright 2017 aixigo AG
+ * Released under the MIT license.
+ * https://laxarjs.org/license
+ */
+
+import { module } from 'angular';
+
+export const name = module( 'finderBoxControl', [] ).name;