+++ application/widgets/open-street-map-widget/messages.js
@@ -1,0 +1,24 @@
+/**
+ * Copyright 2017 aixigo AG
+ * Released under the MIT license.
+ * https://laxarjs.org/license
+ */
+
+export const messages = {
+
+   NO_RESULTS: {
+      de: 'Zu Ihrer Suche konnte leider nichts gefunden werden.',
+      en: 'The request didn\'t return any results.'
+   },
+
+   NO_SEARCH_STARTED: {
+      de: 'Keine Suche gestartet',
+      en: 'No search started'
+   },
+
+   SEARCHING: {
+      de: 'Suche läuft',
+      en: 'Searching ...'
+   }
+
+};