+++ application/widgets/search-box-widget/messages.js
@@ -1,0 +1,14 @@
+/**
+ * Copyright 2017 aixigo AG
+ * Released under the MIT license.
+ * https://laxarjs.org/license
+ */
+
+export const messages = {
+
+   SEARCH_PLACEHOLDER: {
+      de: 'Suche etwas Großartiges',
+      en: 'Search for something awesome'
+   }
+
+};