+++ application/widgets/wikipedia-search-widget/messages.js
@@ -1,0 +1,29 @@
+/**
+ * Copyright 2015 aixigo AG
+ * Released under the MIT license.
+ * http://laxarjs.org/license
+ */
+
+export const messages = {
+
+   EXTERNAL_LINK: {
+      de: 'Zum Artikel',
+      en: 'Open Article'
+   },
+
+   NO_RESULTS: {
+      de: 'Zu Ihrer Suche konnte leider nichts gefunden werden.',
+      en: 'The request didn\'t return any results.'
+   },
+
+   NO_SEARCH_STARTED: {
+      de: 'Keine Suche gestartet',
+      en: 'No search started'
+   },
+
+   SEARCHING: {
+      de: 'Suche läuft',
+      en: 'Searching ...'
+   }
+
+};